+++ PanelFeedbackRepositoryService/src/main/resources/application.properties
@@ -4,7 +4,7 @@ spring.h2.console.enabled=true
 spring.datasource.platform=h2
 spring.datasource.url=jdbc:h2:mem:testdb;MODE=Oracle
 
-spring.rabbitmq.host=ec2-35-168-22-185.compute-1.amazonaws.com
+spring.rabbitmq.host=ec2-34-227-76-119.compute-1.amazonaws.com
 spring.rabbitmq.port=5672
-spring.rabbitmq.username=user
-spring.rabbitmq.password=user
+spring.rabbitmq.username=guest
+spring.rabbitmq.password=guest
